LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.draw.space.basic.worker-expected.txt


PASS U+0020 is rendered the correct size (1em wide) 
